fix: apply clippy suggestions and fix test compilation
- Fix unnecessary type casting in admin.rs (MIN_PAGINATION_LIMIT)
- Use RangeInclusive::contains() instead of manual range checks in voice.rs
- Fix test compilation error: is_str() -> is_string()

+++ src/web/routes/admin.rs
@@ -677,7 +677,7 @@ pub async fn get_users(
         .get("limit")
         .and_then(|v| v.parse().ok())
         .unwrap_or(100)
-        .clamp(MIN_PAGINATION_LIMIT as i64, MAX_PAGINATION_LIMIT);
+        .clamp(MIN_PAGINATION_LIMIT, MAX_PAGINATION_LIMIT);
     let offset = params
         .get("offset")
         .and_then(|v| v.parse().ok())
@@ -887,7 +887,7 @@ pub async fn get_rooms(
         .get("limit")
         .and_then(|v| v.parse().ok())
         .unwrap_or(100)
-        .clamp(MIN_PAGINATION_LIMIT as i64, MAX_PAGINATION_LIMIT);
+        .clamp(MIN_PAGINATION_LIMIT, MAX_PAGINATION_LIMIT);
     let offset = params
         .get("offset")
         .and_then(|v| v.parse().ok())
